Change 9.0 to just 9 and add peridot repos
To reduce commits (every 6 months), this change makes it so "9" is the
absolute default location as it should be symlinked to the latest in
either staging or prod. With that being said, peridot repositories are
also provided for automation purposes but also at the convenience of the
user if they so choose to use those repos instead.

And the readme was updated a bit.